imp: review the layout input and methods ways
- using joystick and adding accelerate button on the right side

+++ Assets/Project/Scripts/Spaceship/Actions/SpaceshipMovementAction.cs
@@ -29,7 +29,7 @@ namespace AsteroidsGame.Actions
 
 #endregion
 
-        private void AccelerateDirection(int direction)
+        private void AccelerateDirection(float direction)
         {
             if(rb.velocity.magnitude >= context.Data.maxForwardVelocity) return;
